+++ src/test/ExSln2/LedBlinker/test_stuff/CArrayMemEx.cs
@@ -1,4 +1,4 @@
-﻿using finlang;
+using finlang;
 
 // test stuff
 namespace ts_;